change size with float k & resize letter & finish L, H, K, U, S

+++ Line.hpp
@@ -15,7 +15,7 @@ void setpos(short int x, short int y);
 // In the CircleArc class, they means the coordinate of center
 class Pattern {
 public:
-    virtual void print(int num1, int num2, int k) const =0;
+    virtual void print(int num1, int num2, double k) const =0;
 protected:
     static char fChar;
 };
@@ -23,7 +23,7 @@ protected:
 class Line: public Pattern {
 public:
     Line(int endX, int endY);
-    virtual void print(int initX, int initY, int k) const =0;
+    virtual void print(int initX, int initY, double k) const =0;
 protected:
     int endX, endY;
 };
@@ -31,36 +31,36 @@ protected:
 class HorizonLine: public Line {
 public:
     HorizonLine(int endX, int endY);
-    void print(int initX, int initY, int k) const;
+    void print(int initX, int initY, double k) const;
 };
 class VerticalLine: public Line {
 public:
     VerticalLine(int endX, int endY);
-    void print (int initX, int initY, int k) const;
+    void print (int initX, int initY, double k) const;
 };
 
 class DiagonalLine: public Line {
 public:
     DiagonalLine(int endX, int endY);
-    void print(int initX, int initY, int k) const;
+    void print(int initX, int initY, double k) const;
 };
 class AntiDiagonalLine: public Line {
 public:
     AntiDiagonalLine(int endX, int endY);
-    void print(int initX, int initY, int k) const;
+    void print(int initX, int initY, double k) const;
 };
 
 class ObliqueLine: public Line {
 public:
     ObliqueLine(int endX, int endY);
-    void print(int initX, int initY, int k) const;
+    void print(int initX, int initY, double k) const;
 private:
     double theta;
 };
 class AntiObliqueLine: public Line {
 public:
     AntiObliqueLine(int endX, int endY);
-    void print(int initX, int initY, int k) const;
+    void print(int initX, int initY, double k) const;
 private:
     double theta;
 };
@@ -68,7 +68,7 @@ private:
 class CircleArc: public Pattern {
 public:
     CircleArc(int begDeg, int endDeg, int rad);
-    void print(int x0, int y0, int k) const;
+    void print(int x0, int y0, double k) const;
 private:
     int begDeg, endDeg, rad;
 };
